add DecodeExt
