+++ crates/fluvio-controlplane-metadata/Cargo.toml
@@ -34,7 +34,7 @@ flv-util = { version = "0.5.0" }
 
 fluvio-types = { version = "0.4.0", path = "../fluvio-types" }
 fluvio-stream-model = { path = "../fluvio-stream-model", version = "0.8.0" }
-fluvio-protocol = { path = "../fluvio-protocol", version = "0.8", features = [
+fluvio-protocol = { path = "../fluvio-protocol", version = "0.9", features = [
     "record",
 ] }
