+++ crates/fluvio-protocol/src/api/response.rs
@@ -1,14 +1,10 @@
-use std::fs::File;
-use std::io::Cursor;
-use std::io::Error as IoError;
-use std::io::ErrorKind;
-use std::io::Read;
-use std::path::Path;
 
-use tracing::debug;
+use std::io::Error as IoError;
+
 use tracing::trace;
 
 use bytes::{Buf, BufMut};
+use crate::DecodeExt;
 use crate::api::RequestHeader;
 use crate::{Decoder, Encoder, Version};
 
@@ -32,13 +28,11 @@ impl<P> ResponseMessage<P> {
     }
 }
 
-impl<P> ResponseMessage<P>
-where
-    P: Decoder,
+impl <P> DecodeExt for ResponseMessage<P> where P: DecodeExt
 {
-    pub fn decode_from<T>(src: &mut T, version: Version) -> Result<Self, IoError>
+    fn decode_from<T>(src: &mut T, version: Version) -> Result<Self, IoError>
     where
-        T: Buf,
+        T: Buf
     {
         let mut correlation_id: i32 = 0;
         correlation_id.decode(src, version)?;
@@ -51,33 +45,7 @@ where
         })
     }
 
-    pub fn decode_from_file<H: AsRef<Path>>(
-        file_name: H,
-        version: Version,
-    ) -> Result<Self, IoError> {
-        debug!("decoding from file: {:#?}", file_name.as_ref());
-        let mut f = File::open(file_name)?;
-        let mut buffer: [u8; 1000] = [0; 1000];
-
-        f.read_exact(&mut buffer)?;
-        let data = buffer.to_vec();
-
-        let mut src = Cursor::new(&data);
-
-        // ResponseMessage implementation of fluvio_protocol::storage::FileWrite trait first encodes the length
-        // of the ResponseMessage
-        let mut size: i32 = 0;
-        size.decode(&mut src, version)?;
-        trace!("decoded response size: {} bytes", size);
-
-        if src.remaining() < size as usize {
-            return Err(IoError::new(
-                ErrorKind::UnexpectedEof,
-                "not enough for response",
-            ));
-        }
-        Self::decode_from(&mut src, version)
-    }
+    
 }
 
 impl<P> Encoder for ResponseMessage<P>
