+++ crates/fluvio-protocol/src/record/data.rs
@@ -517,7 +517,8 @@ where
 
 impl<B> Decoder for Record<B>
 where
-    B: Decoder,
+    Option<B>: Decoder,
+    B: Decoder
 {
     fn decode<T>(&mut self, src: &mut T, version: Version) -> Result<(), Error>
     where
