+++ crates/fluvio-smartengine/Cargo.toml
@@ -32,7 +32,7 @@ cfg-if = "1.0.0"
 derive_builder = { workspace = true }
 
 fluvio-future = { version = "0.4.0", default-features = false }
-fluvio-protocol = { path = "../fluvio-protocol", version = "0.8.0", features = [
+fluvio-protocol = { path = "../fluvio-protocol", version = "0.9.0", features = [
     "record",
 ] }
 fluvio-smartmodule = { path = "../fluvio-smartmodule", version = "0.4.0", default-features = false }
