+++ crates/fluvio-spu-schema/Cargo.toml
@@ -27,7 +27,7 @@ flate2 = { version = "1.0.20" }
 fluvio-types = { version = "0.4.0", path = "../fluvio-types" }
 fluvio-future = { version = "0.4.0", optional = true }
 # Fluvio dependencies
-fluvio-protocol = { path = "../fluvio-protocol", version = "0.8.0", features = [
+fluvio-protocol = { path = "../fluvio-protocol", version = "0.9.0", features = [
     "api",
     "record",
 ] }
